+++ Editor/Scripts/Typed/FloatIntGameEventInspector.cs
@@ -1,0 +1,8 @@
+using UnityEditor;
+
+namespace FredericRP.EventManagement
+{
+  [CustomEditor(typeof(FloatIntGameEvent))]
+  public class FloatIntGameEventInspector : TwoTypeGameEventInspector<float, int>
+  { }
+}