+++ Editor/Scripts/Typed/FloatStringGameEventInspector.cs
@@ -1,0 +1,8 @@
+using UnityEditor;
+
+namespace FredericRP.EventManagement
+{
+  [CustomEditor(typeof(FloatStringGameEvent))]
+  public class FloatStringGameEventInspector : TwoTypeGameEventInspector<float, string>
+  { }
+}