+++ Editor/Scripts/Typed/StringFloatGameEventInspector.cs
@@ -1,0 +1,8 @@
+using UnityEditor;
+
+namespace FredericRP.EventManagement
+{
+  [CustomEditor(typeof(StringFloatGameEvent))]
+  public class StringFloatGameEventInspector : TwoTypeGameEventInspector<string, float>
+  { }
+}