+++ Editor/Scripts/Typed/StringIntGameEventInspector.cs
@@ -2,8 +2,7 @@ using UnityEditor;
 
 namespace FredericRP.EventManagement
 {
-  [CustomEditor(typeof(StringIntGameEvent))]
-  public class StringIntGameEventInspector : TwoTypeGameEventInspector<string, int>
-  {
-  }
+	[CustomEditor(typeof(StringIntGameEvent))]
+	public class StringIntGameEventInspector : TwoTypeGameEventInspector<string, int>
+	{ }
 }