+++ Runtime/Scripts/TypedGameEvent/Typed/FloatIntGameEvent.cs
@@ -1,0 +1,8 @@
+using UnityEngine;
+
+namespace FredericRP.EventManagement
+{
+  [CreateAssetMenu(menuName = "FredericRP/Float-Int Game Event")]
+  public class FloatIntGameEvent : TwoTypeGameEvent<float, int>
+  { }
+}