+++ Runtime/Scripts/TypedGameEvent/Typed/FloatStringGameEvent.cs
@@ -1,0 +1,8 @@
+using UnityEngine;
+
+namespace FredericRP.EventManagement
+{
+  [CreateAssetMenu(menuName = "FredericRP/Float-String Game Event")]
+  public class FloatStringGameEvent : TwoTypeGameEvent<float, string>
+  { }
+}