+++ Runtime/Scripts/TypedGameEvent/Typed/StringFloatGameEvent.cs
@@ -1,0 +1,8 @@
+using UnityEngine;
+
+namespace FredericRP.EventManagement
+{
+  [CreateAssetMenu(menuName = "FredericRP/String-Float Game Event")]
+  public class StringFloatGameEvent : TwoTypeGameEvent<string, float>
+  { }
+}