+++ Runtime/Scripts/TypedGameEvent/Typed/StringIntGameEvent.cs
@@ -2,8 +2,7 @@ using UnityEngine;
 
 namespace FredericRP.EventManagement
 {
-  [CreateAssetMenu(menuName = "FredericRP/String-Int Game Event")]
-  public class StringIntGameEvent : TwoTypeGameEvent<string, int>
-  {
-  }
+	[CreateAssetMenu(menuName = "FredericRP/String-Int Game Event")]
+	public class StringIntGameEvent : TwoTypeGameEvent<string, int>
+	{ }
 }